// src/app/models/cartItem.ts
import { Product } from './product';

export class CartItem {
    quantity: number;
    productModel: Product;
//   opt1: Option;
//  opt2: Option;
//   opt3: Option;
//   opt4: Option;
//   opt5: Option;
//   opt6: Option;
//   opt7: Option;
//   opt8: Option;
//   opt9: Option;
//   opt10: Option;
//   opt11: Option;
//   opt12: Option;
//   opt13: Option;
//   opt14: Option;
//   opt15: Option;
//   opt16?: Option;
//   opt17?: Option;
  
}